# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.3
info:
  title: B-ting Backend API
  description: B-ting backend API documentation.
  version: 0.0.1-SNAPSHOT
servers:
  - url: https://api.buting.store
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/chat/rooms/zone:
    get:
      tags:
        - Chat
      summary: 권역별 오픈채팅방 목록 조회
      description: 선택한 행정 권역(ChatZone)에 개설되어 있는 오픈채팅방 리스트를 조회합니다.
      parameters:
        - name: zone
          in: query
          required: true
          description: 조회할 부산 지역 권역 코드 (이넘값)
          schema:
            type: string
            enum:
              - SUYEONG_NAMGU
              - HAEUNDAE_GIJANG
              - YEONGDO
              - CENTRAL_NORTH
              - OLD_DOWNTOWN
              - WESTERN_BUSAN
            example: SUYEONG_NAMGU
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocalChatroomResponse"
        "400":
          description: Validation failed (권역 정보 누락 등)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/chat/rooms/{roomId}/join:
    post:
      tags:
        - Chat
      summary: 특정 오픈채팅방 가입하기
      description: 특정 오픈채팅방을 가입하기, 방인원수는 1증가
      security:
        - opaqueToken: []
      parameters:
        - name: roomId
          in: path
          required: true
          description: 가입할 채팅방 UUID
          schema:
            type: string
            format: uuid
            example: cf7cabe7-68ef-4296-a33c-9a912d708734
      responses:
        "200":
          description: 채팅방 가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "채팅방 가입 완료"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/chat/rooms/{roomId}/exit:
    post:
      tags:
        - Chat
      summary: 특정 오픈채팅방 퇴장 (나가기)
      description: 참여 중이던 오픈채팅방에서 나갑니다. 매핑 데이터는 Hard Delete되며, 방 인원수가 1 차감됩니다.
      security:
        - opaqueToken: []
      parameters:
        - name: roomId
          in: path
          required: true
          description: 퇴장할 채팅방 UUID
          schema:
            type: string
            format: uuid
            example: cf7cabe7-68ef-4296-a33c-9a912d708734
      responses:
        "200":
          description: 채팅방 퇴장 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "채팅방 퇴장이 완료되었습니다."
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/chat/rooms/{roomId}/messages:
    get:
      tags:
        - Chat
      summary: 채팅방 과거 대화 내역 조회 (커서 기반 페이징)
      description: |
        사용자가 오픈채팅방에 진입하거나 스크롤을 위로 올려 과거 내역을 더 불러올 때 호출합니다.

        - **첫 진입 시**: `lastMessageId`를 비우고 요청하면 가장 최신 메시지 100개를 과거->최신(오름차순) 순으로 정렬하여 반환합니다.
        - **더보기 요청 시**: 현재 클라이언트가 가진 메시지 중 가장 오래된 메시지의 `UUID`를 `lastMessageId`로 실어서 보내면, 해당 메시지 이전의 대화 내역 100개를 반환합니다.
      security:
        - opaqueToken: [ ]
      parameters:
        - name: roomId
          in: path
          required: true
          description: 내역을 조회할 채팅방 UUID
          schema:
            type: string
            format: uuid
            example: cf7cabe7-68ef-4296-a33c-9a912d708734
        - name: lastMessageId
          in: query
          required: false
          description: |
            더보기 페이징용 커서 ID. 
            현재 화면에 보이는 가장 오래된(상단) 메시지의 messageId를 전달합니다. (첫 진입 시 생략)
          schema:
            type: string
            format: uuid
            example: 52a58863-2360-4950-862f-2ba410a9d422
      responses:
        "200":
          description: 과거 대화 내역 조회 성공 (최대 100건, 과거 -> 최신순 정렬 배열)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageListEnvelope"
        "400":
          description: 존재하지 않는 방 ID, 존재하지 않는 기준 메시지 ID(커서)이거나 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized (만료되거나 누락된 Opaque 토큰)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
...
  securitySchemes:
    opaqueToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Email is required.
        data:
          nullable: true
...
    LocalChatroomResponse:
      type: object
      properties:
        roomId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        title:
          type: string
          example: "수영구/남구 친목 채팅방"
        description:
          type: string
          nullable: true
          example: "수영구와 남구 주민들을 위한 소통 공간입니다."
        chatZone:
          type: string
          enum:
            - SUYEONG_NAMGU
            - HAEUNDAE_GIJANG
            - JUNG_YEONGDO
          example: SUYEONG_NAMGU
        maxMembers:
          type: integer
          example: 100
        currentMembers:
          type: integer
          example: 14
        updatedAt:
          type: string
          format: date-time
          example: "2026-06-30T10:47:06Z"
    ChatMessageResponse:
      type: object
      properties:
        messageId:
          type: string
          format: uuid
          description: 메시지 고유 식별자 ID (uuid)
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa3
        roomId:
          type: string
          format: uuid
          description: 채팅방 고유 UUID
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        userId:
          type: string
          format: uuid
          description: 메시지를 보낸 유저 고유 UUID
          example: d1234567-89ab-cdef-0123-456789abcdef
        senderNickname:
          type: string
          description: 보낸 유저의 닉네임
          example: "광안리보이"
        content:
          type: string
          description: 채팅 메시지 본문 내용
          example: "안녕하세요! 지난 대화들 확인하고 들어왔습니다."
        createdAt:
          type: string
          format: date-time
          description: 메시지 생성 일시 (ISO 8601 포맷)
          example: "2026-07-01T23:15:00Z"
        isMine:
          type: boolean
          nullable: true
          description: REST 히스토리 조회에서 현재 사용자가 보낸 메시지 여부입니다. STOMP 브로드캐스트에서는 null일 수 있습니다.
          example: false
    ChatMessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessageResponse"
    ChatMessageListEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 채팅 메시지 목록 조회
        data:
          $ref: "#/components/schemas/ChatMessageListResponse"